ImageColorPalette/requirements.txt
pandas==2.2.2
Flask==3.0.2
pillow==10.2.0
Werkzeug==3.0.1